chore: move notes related hooks
